Move HelperMethods module to separate file
