+++ lib/uri_format_validator.rb
@@ -1,6 +1,7 @@
 require "uri_format_validator/version"
 require "uri_format_validator/localization"
 require "validators/uri_format_validator"
+require "validators/helper_methods"
 
 # This is a placeholder...
 module UriFormatValidator
